fix(roles): flag isPro
